MVP pass plus some extra

+++ final-client/src/components/VaultComponent.vue
@@ -1,16 +1,18 @@
 <template>
-  <div class="vault-component col-3">
+  <div class="vault-component">
     <div class="card">
-      <div class="card-title">
-        <div v-if="state.account">
-          <h3>{{ vault.name }} <i class="fa fa-trash" aria-hidden="true" @click="deleteVault" v-if="state.account.id === vault.creatorId"></i></h3>
+      <div class="card-container">
+        <div class="card-title">
+          <div v-if="state.account">
+            <h3>{{ vault.name }} <span><i class="fa fa-trash" aria-hidden="true" @click="deleteVault" v-if="state.account.id === vault.creatorId"></i></span> </h3>
+          </div>
         </div>
+        <router-link :to="{name: 'VaultPage', params: {id: vault.id}}">
+          <div class="card-text">
+            {{ vault.description }}
+          </div>
+        </router-link>
       </div>
-      <router-link :to="{name: 'VaultPage', params: {id: vault.id}}">
-        <div class="card-text">
-          {{ vault.description }}
-        </div>
-      </router-link>
     </div>
   </div>
 </template>
@@ -36,7 +38,10 @@ export default {
       state,
       route,
       deleteVault() {
-        vaultService.deleteVault(props.vault.id, route.params.id)
+        const res = window.confirm('Are you sure you want to delete this?')
+        if (res) {
+          vaultService.deleteVault(props.vault.id, route.params.id)
+        }
       }
     }
   },
@@ -45,9 +50,5 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-
-.card {
-  height: 100%;
-}
 
 </style>
